moved external registry to pytest-simcore
